Refactor formatting visitor to use java reflections
This will improve formatting visitor to use java reflections
to visit through each nodes available in the AST.
