+++ composer/modules/web/source-gen/SourceGenTemplate.java
@@ -44,7 +44,7 @@ public class SourceGen {
 
 
     @FindbugsSuppressWarnings
-    String getSourceOf(JsonObject node, boolean pretty, boolean replaceLambda) {
+    public String getSourceOf(JsonObject node, boolean pretty, boolean replaceLambda) {
         if (node == null) {
             return "";
         }
