+++ language-server/modules/langserver-core/src/test/java/org/ballerinalang/langserver/common/util/CommonUtil.java
@@ -20,6 +20,8 @@ package org.ballerinalang.langserver.common.util;
 import com.google.gson.Gson;
 import org.ballerinalang.langserver.BallerinaLanguageServer;
 import org.eclipse.lsp4j.DidOpenTextDocumentParams;
+import org.eclipse.lsp4j.DocumentFormattingParams;
+import org.eclipse.lsp4j.FormattingOptions;
 import org.eclipse.lsp4j.Position;
 import org.eclipse.lsp4j.ReferenceContext;
 import org.eclipse.lsp4j.ReferenceParams;
@@ -95,6 +97,28 @@ public class CommonUtil {
 
                 serviceEndpoint.notify("textDocument/didOpen", didOpenTextDocumentParams);
                 result = serviceEndpoint.request(method, referenceParams);
+                break;
+            case "textDocument/formatting":
+                DocumentFormattingParams documentFormattingParams = new DocumentFormattingParams();
+
+                TextDocumentIdentifier textDocumentIdentifier1 = new TextDocumentIdentifier();
+                textDocumentIdentifier1.setUri(Paths.get(file).toUri().toString());
+
+                FormattingOptions formattingOptions = new FormattingOptions();
+                formattingOptions.setInsertSpaces(true);
+                formattingOptions.setTabSize(4);
+
+                documentFormattingParams.setOptions(formattingOptions);
+                documentFormattingParams.setTextDocument(textDocumentIdentifier1);
+
+                DidOpenTextDocumentParams didOpenTextDocumentParams1 = new DidOpenTextDocumentParams();
+                TextDocumentItem textDocument1 = new TextDocumentItem();
+                textDocument1.setUri(textDocumentIdentifier1.getUri());
+                textDocument1.setText(fileContent);
+                didOpenTextDocumentParams1.setTextDocument(textDocument1);
+
+                serviceEndpoint.notify("textDocument/didOpen", didOpenTextDocumentParams1);
+                result = serviceEndpoint.request(method, documentFormattingParams);
                 break;
             default:
                 break;
